# docs/endpoints/openapi.yaml
openapi: 3.0.3
info:
  title: GymSheet API
  version: 1.1.0-hardening
  description: |
    API autenticada para perfiles, catálogo de ejercicios, multimedia y sesiones de entrenamiento.
    Los nombres externos en español se mantienen por compatibilidad con la versión 1.
servers:
  - url: http://localhost:3000/api/v1
    description: Desarrollo local
security:
  - bearerAuth: []
...
paths:
...
  /exercises:
    get:
      tags: [Exercises]
      summary: Listar ejercicios visibles
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSize25Param'
        - name: search
          in: query
          schema: { type: string, minLength: 1, maxLength: 120 }
        - name: grupoMuscular
          in: query
          schema: { type: string, maxLength: 100 }
        - name: equipoId
          in: query
          schema: { type: string, format: uuid }
        - name: bodyPart
          in: query
          schema: { type: string, maxLength: 100 }
        - name: targetMuscle
          in: query
          schema: { type: string, maxLength: 120 }
        - name: dataSource
          in: query
          schema:
            type: string
            enum: [CUSTOM, EXERCISES_DATASET]
      responses:
        '200':
          description: Página de ejercicios
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExercisePageEnvelope'
        '400': { $ref: '#/components/responses/Problem400' }
  /exercises/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [Exercises]
      summary: Obtener ejercicio visible
      responses:
        '200':
          description: Ejercicio visible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExerciseEnvelope'
        '404': { $ref: '#/components/responses/Problem404' }
    patch:
      tags: [Exercises]
      summary: Actualizar ejercicio personal propio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateExerciseRequest'
      responses:
        '200':
          description: Ejercicio actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExerciseEnvelope'
        '400': { $ref: '#/components/responses/Problem400' }
        '403': { $ref: '#/components/responses/Problem403' }
        '404': { $ref: '#/components/responses/Problem404' }
    delete:
      tags: [Exercises]
      summary: Inhabilitar ejercicio personal propio
      responses:
        '200':
          description: Ejercicio inhabilitado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExerciseEnvelope'
        '403': { $ref: '#/components/responses/Problem403' }
        '404': { $ref: '#/components/responses/Problem404' }
  /exercises/personal:
    post:
      tags: [Exercises]
      summary: Crear ejercicio personal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExerciseRequest'
      responses:
        '201':
          description: Ejercicio creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExerciseEnvelope'
        '400': { $ref: '#/components/responses/Problem400' }
  /admin/exercises/global:
    post:
      tags: [Administration]
      summary: Crear ejercicio global
      x-required-role: ADMIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExerciseRequest'
      responses:
        '201':
          description: Ejercicio global creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExerciseEnvelope'
        '400': { $ref: '#/components/responses/Problem400' }
        '403': { $ref: '#/components/responses/Problem403' }
  /admin/exercises/global/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    patch:
      tags: [Administration]
      summary: Actualizar ejercicio global
      x-required-role: ADMIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateExerciseRequest'
      responses:
        '200':
          description: Ejercicio global actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExerciseEnvelope'
        '400': { $ref: '#/components/responses/Problem400' }
        '403': { $ref: '#/components/responses/Problem403' }
        '404': { $ref: '#/components/responses/Problem404' }
    delete:
      tags: [Administration]
      summary: Inhabilitar ejercicio global
      x-required-role: ADMIN
      responses:
        '200':
          description: Ejercicio global inhabilitado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExerciseEnvelope'
        '403': { $ref: '#/components/responses/Problem403' }
        '404': { $ref: '#/components/responses/Problem404' }
  /exercises/{exerciseId}/media:
    parameters:
      - name: exerciseId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Exercise media]
      summary: Listar multimedia activa del ejercicio visible
      responses:
        '200':
          description: Multimedia activa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExerciseMediaListEnvelope'
        '404': { $ref: '#/components/responses/Problem404' }
    post:
      tags: [Exercise media]
      summary: Registrar multimedia
      description: Requiere propiedad del ejercicio personal o rol ADMIN para un ejercicio global.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExerciseMediaRequest'
      responses:
        '201':
          description: Multimedia registrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExerciseMediaEnvelope'
        '400': { $ref: '#/components/responses/Problem400' }
        '403': { $ref: '#/components/responses/Problem403' }
        '404': { $ref: '#/components/responses/Problem404' }
        '409': { $ref: '#/components/responses/Problem409' }
  /exercise-media/{mediaId}:
    parameters:
      - name: mediaId
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      tags: [Exercise media]
      summary: Inhabilitar multimedia
      responses:
        '200':
          description: Multimedia inhabilitada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteEnvelope'
        '403': { $ref: '#/components/responses/Problem403' }
        '404': { $ref: '#/components/responses/Problem404' }
...
  /user-exercises/{exerciseId}:
    parameters:
      - name: exerciseId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Favorites]
      summary: Agregar ejercicio frecuente
      responses:
        '201':
          description: Frecuente creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObjectEnvelope'
        '404': { $ref: '#/components/responses/Problem404' }
        '409': { $ref: '#/components/responses/Problem409' }
    delete:
      tags: [Favorites]
      summary: Quitar ejercicio frecuente
      responses:
        '200':
          description: Frecuente eliminado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteEnvelope'
        '404': { $ref: '#/components/responses/Problem404' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize25Param:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
...
  responses:
    Problem400:
      description: Solicitud inválida
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    Problem403:
      description: Operación no autorizada
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Problem404:
      description: Recurso no encontrado o no visible
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Problem409:
      description: Conflicto de estado, identidad o regla de negocio
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    Problem:
      type: object
      required: [type, title, status, detail, instance, requestId]
      properties:
        type: { type: string, format: uri-reference, example: about:blank }
        title: { type: string, example: Bad Request }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string }
        requestId: { type: string, format: uuid }
        errors:
          type: array
          items:
            type: object
            additionalProperties: true
    SuccessEnvelope:
      type: object
      required: [ok, data]
      properties:
        ok: { type: boolean, enum: [true] }
        data: {}
    GenericObjectEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              type: object
              additionalProperties: true
    DeleteEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              type: object
              required: [deleted]
              properties:
                deleted: { type: boolean, enum: [true] }
...
    LocalizedInstructions:
      type: object
      maxProperties: 20
      additionalProperties:
        type: string
        minLength: 1
        maxLength: 6000
    LocalizedInstructionSteps:
      type: object
      maxProperties: 20
      additionalProperties:
        type: array
        minItems: 1
        maxItems: 30
        items: { type: string, minLength: 1, maxLength: 1000 }
    CreateExerciseRequest:
      type: object
      required: [nombre, grupoMuscular]
      additionalProperties: false
      properties:
        nombre: { type: string, minLength: 2, maxLength: 160 }
        grupoMuscular: { type: string, minLength: 2, maxLength: 120 }
        descripcion: { type: string, nullable: true, maxLength: 2000 }
        equipoIds:
          type: array
          maxItems: 30
          items: { type: string, format: uuid }
        bodyPart: { type: string, nullable: true, minLength: 2, maxLength: 100 }
        targetMuscle: { type: string, nullable: true, minLength: 2, maxLength: 120 }
        synergistMuscleGroup: { type: string, nullable: true, minLength: 2, maxLength: 120 }
        secondaryMuscles:
          type: array
          maxItems: 30
          items: { type: string, minLength: 2, maxLength: 120 }
        instructions: { $ref: '#/components/schemas/LocalizedInstructions' }
        instructionSteps: { $ref: '#/components/schemas/LocalizedInstructionSteps' }
        metadata:
          type: object
          description: El JSON serializado no puede superar 16 KiB.
          additionalProperties: true
    UpdateExerciseRequest:
      type: object
      minProperties: 1
      additionalProperties: false
      properties:
        nombre: { type: string, minLength: 2, maxLength: 160 }
        grupoMuscular: { type: string, minLength: 2, maxLength: 120 }
        descripcion: { type: string, nullable: true, maxLength: 2000 }
        equipoIds:
          type: array
          maxItems: 30
          items: { type: string, format: uuid }
        bodyPart: { type: string, nullable: true, minLength: 2, maxLength: 100 }
        targetMuscle: { type: string, nullable: true, minLength: 2, maxLength: 120 }
        synergistMuscleGroup: { type: string, nullable: true, minLength: 2, maxLength: 120 }
        secondaryMuscles:
          type: array
          maxItems: 30
          items: { type: string, minLength: 2, maxLength: 120 }
        instructions: { $ref: '#/components/schemas/LocalizedInstructions' }
        instructionSteps: { $ref: '#/components/schemas/LocalizedInstructionSteps' }
        metadata:
          type: object
          additionalProperties: true
    Exercise:
      type: object
      required: [id, nombre, grupoMuscular, tipo, estado]
      additionalProperties: true
      properties:
        id: { type: string, format: uuid }
        nombre: { type: string }
        grupoMuscular: { type: string }
        descripcion: { type: string, nullable: true }
        tipo: { type: string, enum: [GLOBAL, PERSONAL] }
        estado: { type: string, enum: [ACTIVO, INACTIVO] }
        dataSource: { type: string, enum: [CUSTOM, EXERCISES_DATASET] }
        media:
          type: array
          items: { $ref: '#/components/schemas/ExerciseMedia' }
    ExerciseEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/Exercise' }
    ExercisePageEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              type: object
              required: [items, page, pageSize, total, totalPages]
              properties:
                items:
                  type: array
                  items: { $ref: '#/components/schemas/Exercise' }
                page: { type: integer, minimum: 1 }
                pageSize: { type: integer, minimum: 1, maximum: 100 }
                total: { type: integer, minimum: 0 }
                totalPages: { type: integer, minimum: 0 }
    CreateExerciseMediaRequest:
      type: object
      required: [mediaType, provider, url, altText]
      additionalProperties: false
      properties:
        mediaType: { type: string, enum: [IMAGE, GIF, VIDEO] }
        provider: { type: string, enum: [EXTERNAL_URL, CLOUDINARY, S3, LOCAL] }
        externalId: { type: string, nullable: true, minLength: 1, maxLength: 180 }
        url: { type: string, format: uri, maxLength: 2048, pattern: '^https://' }
        thumbnailUrl: { type: string, format: uri, nullable: true, maxLength: 2048, pattern: '^https://' }
        mimeType: { type: string, nullable: true, minLength: 3, maxLength: 120 }
        width: { type: integer, nullable: true, minimum: 1, maximum: 10000 }
        height: { type: integer, nullable: true, minimum: 1, maximum: 10000 }
        checksumSha256: { type: string, nullable: true, pattern: '^[a-fA-F0-9]{64}$' }
        altText: { type: string, minLength: 3, maxLength: 500 }
        attribution: { type: string, nullable: true, maxLength: 2000 }
        license: { type: string, nullable: true, maxLength: 160 }
        isPrimary: { type: boolean, default: false }
        sortOrder: { type: integer, minimum: 0, maximum: 1000, default: 0 }
        metadata:
          type: object
          additionalProperties: true
          default: {}
    ExerciseMedia:
      type: object
      required: [id, exerciseId, mediaType, provider, url, altText, isPrimary, sortOrder, status]
      additionalProperties: true
      properties:
        id: { type: string, format: uuid }
        exerciseId: { type: string, format: uuid }
        mediaType: { type: string, enum: [IMAGE, GIF, VIDEO] }
        provider: { type: string, enum: [EXTERNAL_URL, CLOUDINARY, S3, LOCAL] }
        url: { type: string, format: uri }
        altText: { type: string }
        isPrimary: { type: boolean }
        sortOrder: { type: integer }
        status: { type: string, enum: [ACTIVE, INACTIVE] }
    ExerciseMediaEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/ExerciseMedia' }
    ExerciseMediaListEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/ExerciseMedia' }
...